# runtime/swagger/http.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MineAdmin API DOC",
        "description": "MineAdmin后台接口列表",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9501",
            "description": "本地服务器"
        }
    ],
    "paths": {
...
        "/system/dept/index": {
            "get": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/recycle": {
            "get": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/tree": {
            "get": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptTreeGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/getLeaderList": {
            "get": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptGetLeaderListGET",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "null"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/save": {
            "post": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/addLeader": {
            "post": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptAddLeaderPOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/delLeader": {
            "delete": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptDelLeaderDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/update/{id}": {
            "put": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/delete": {
            "delete": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/realDelete": {
            "delete": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/recovery": {
            "put": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/changeStatus": {
            "put": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/numberOperation": {
            "put": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptNumberOperationPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/dept/remote": {
            "post": {
                "tags": [
                    "DeptController"
                ],
                "summary": "",
                "operationId": "SystemDeptRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/index": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/recycle": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/save": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/update/{id}": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/delete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/realDelete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/recovery": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/changeStatus": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/remote": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}